backend/requirements.txt
# Flask Backend Requirements for PixelFly - Production Ready
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Production WSGI Servers
gunicorn==21.2.0  # For Linux/macOS
waitress==2.1.2   # For Windows

# Core Image Processing (Vercel Compatible)
Pillow==10.0.1
numpy==1.24.3

# HTTP Requests
requests==2.31.0

# Note: Heavy libraries like opencv-python, scikit-image, and AI libraries
# are commented out for Vercel deployment to avoid size limits
# Uncomment if deploying to a different platform with more resources

# opencv-python==4.8.1.78
# google-generativeai==0.3.2
# langchain==0.0.335
# scikit-image==0.21.0
